fix(gemma4): close routed experts review gaps
