+++ docs/models/gemma4/serving.md
@@ -1,6 +1,6 @@
 # Gemma 4 serving
 
-**TL;DR:** The engine schedules per iteration: up to the configured decode slots (16 by default) hold requests, each prompt prefills whole at a step boundary by default, and every active request advances one token per batched step. Prompt plus output past the ceiling — 8192 by default, raised up to the checkpoint's 262144 by `PEGAINFER_MAX_CONTEXT` — is refused at admission, while a request that only has to wait for a decode slot queues instead. The two KV families are budgeted separately (7.27 GiB sliding + 2.00 GiB global at 12B, at the defaults). **The default configuration needs a 48 GiB card**: it sits at 32.2 GiB before it serves anything, so a 32 GiB device cannot start it — smaller envelopes are a matter of the slots and ceiling knobs below. A row's output moves with the bucket widths it decodes at, but not with what its companions contain. An opt-in conversation prefix cache (`PEGAINFER_PREFIX_CACHE=K`) resumes multi-turn prompts at the cost of a pre-allocated page budget, and an opt-in overlap lane (`PEGAINFER_ASYNC_PREFILL=green:NN`) trades prefill latency for decode-tail protection under long-prompt admissions. An opt-in chunked walk (`PEGAINFER_MIX_CHUNK_TOKENS=N`) bounds how many prompt rows a mixed admission computes per step, so live streams advance per segment instead of waiting out whole prompts, and a raised ceiling (`PEGAINFER_MAX_CONTEXT`, with `PEGAINFER_DECODE_SLOTS` trading concurrency for context) serves long-context workloads on the same card.
+**TL;DR:** The engine schedules per iteration: up to the configured decode slots (16 by default) hold requests, each prompt prefills whole at a step boundary by default, and every active request advances one token per batched step. Prompt plus output past the ceiling — 8192 by default, raised up to the checkpoint's 262144 by `PEGAINFER_MAX_CONTEXT` — is refused at admission, while a request that only has to wait for a decode slot queues instead. The two KV families are budgeted separately (7.27 GiB sliding + 2.00 GiB global at 12B, at the defaults). **The default configuration needs a 48 GiB card**: it sits at 32.2 GiB before it serves anything, so a 32 GiB device cannot start it — smaller envelopes are a matter of the slots and ceiling knobs below. A row's output moves with the bucket widths it decodes at, but not with what its companions contain. An opt-in conversation prefix cache (`PEGAINFER_PREFIX_CACHE=K`) resumes multi-turn prompts at the cost of a pre-allocated page budget, and a dense-only opt-in overlap lane (`PEGAINFER_ASYNC_PREFILL=green:NN`) trades prefill latency for decode-tail protection under long-prompt admissions. An opt-in chunked walk (`PEGAINFER_MIX_CHUNK_TOKENS=N`) bounds how many prompt rows a mixed admission computes per step, so live streams advance per segment instead of waiting out whole prompts, and a raised ceiling (`PEGAINFER_MAX_CONTEXT`, with `PEGAINFER_DECODE_SLOTS` trading concurrency for context) serves long-context workloads on the same card. The routed 26B is eager-only.
 
 Last touched: 2026-08
 
@@ -92,7 +92,7 @@ At `PEGAINFER_PREFIX_CACHE=16` the idle footprint measured **39242 MiB** against
 
 ## The async prefill lane (opt-in)
 
-`PEGAINFER_ASYNC_PREFILL` (unset by default) moves a live-batch admission's prefill onto its own stream, so decode steps keep replaying while the prompt computes. `green:NN` pins the lane to roughly NN% of the SMs via a Green Context — the cap is the mechanism: a `shared` lane's full-width prefill grids starve decode steps, and is kept only for comparison. An unrecognized value or an unviable SM partition refuses to start rather than silently degrading.
+`PEGAINFER_ASYNC_PREFILL` is supported by dense checkpoints only. A routed checkpoint refuses any configured value. When unset, dense serving uses the normal mixed-step path; when set, a live-batch admission's prefill moves onto its own stream so decode steps keep replaying while the prompt computes. `green:NN` pins the lane to roughly NN% of the SMs via a Green Context — the cap is the mechanism: a `shared` lane's full-width prefill grids starve decode steps, and is kept only for comparison. An unrecognized value or an unviable SM partition refuses to start rather than silently degrading.
 
 One prefill is in flight at most; further arrivals wait while decode keeps stepping, a prompt arriving with nothing active takes the sync path, a restored prefix-cache hit prefills only its unseen suffix, and the sliding window's front release is deferred to the join so no page can be re-allocated under in-flight reads.
 
@@ -138,7 +138,7 @@ A row decoding in a batch does not produce the same logprobs as the same row dec
 
 Hold the trajectory fixed and replace what the other rows are — their content, their prompt lengths — and the row is bit-identical, so **no companion row contaminates it**. Change when companions arrive or retire and the row moves, because the kernels pick shapes and reduction orders by batch size. (That a row reads the *right* positions and page rows in the first place is a separate question, gated by the preps' closed-form tests rather than by this comparison.)
 
-Decode steps compute at power-of-two batch buckets — a batch pads to its bucket with rows that write the pools' reserved padding pages — and are replayed as per-bucket CUDA graphs captured at startup (`--cuda-graph=false` is the eager escape hatch; padding applies either way, so the two modes are the same arithmetic). Bucketing also quantizes the width trajectory: batch sizes that share a bucket share their arithmetic.
+For dense checkpoints, decode steps compute at power-of-two batch buckets — a batch pads to its bucket with rows that write the pools' reserved padding pages — and are replayed as per-bucket CUDA graphs captured at startup (`--cuda-graph=false` is the eager escape hatch; padding applies either way, so the two modes are the same arithmetic). Routed checkpoints always use eager execution. Their graph experiment improved a single stream by 6%, but four concurrent streams ranged from 1.4 to 9.1 seconds as shrinking batches crossed buckets and recaptured, so the runtime logs and disables graph capture for them. Bucketing also quantizes the width trajectory: batch sizes that share a bucket share their arithmetic.
 
 The consequence for callers: **greedy output is reproducible for a given workload on an otherwise idle device, not across workloads.** Replaying the same requests the same way returns the same tokens; sending them alongside different traffic changes the widths they decode at and can flip a near-tie. Another process on the same GPU does this too, by moving when each prompt's prefill lands relative to the decodes around it.
