+++ pegainfer-kernels/csrc/kimi_k2/kimi_experts.cu
@@ -1,14 +1,8 @@
 #include "../common.cuh"
+#include "../marlin/ffi.cuh"
 
 #include <cuda.h>
 #include <stdint.h>
-
-CUresult marlin_moe_align_block_size_cuda(
-    const int* topk_idx, int* sorted_token_ids, int* expert_ids,
-    int* num_tokens_post_padded, uint32_t* expert_offsets,
-    uint32_t* expert_cursor, int active_tokens, int topk, int global_start,
-    int local_experts, int block_size, int max_padded_tokens, int max_m_blocks,
-    cudaStream_t stream);
 
 extern "C" {
